# backend/internal/docs/swagger.yaml
definitions:
  handler.ChangePasswordRequest:
    properties:
      new_password:
        maxLength: 128
        minLength: 6
        type: string
      old_password:
        type: string
      username:
        type: string
    required:
    - new_password
    - old_password
    - username
    type: object
  handler.LoginRequest:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
...
  handler.OAuth2TokenResponse:
    properties:
      access_token:
        type: string
      expires_in:
        type: integer
      token_type:
        type: string
    type: object
...
  handler.RefreshRequest:
    properties:
      refresh_token:
        type: string
    required:
    - refresh_token
    type: object
  handler.RegisterRequest:
    properties:
      password:
        maxLength: 128
        minLength: 6
        type: string
      username:
        maxLength: 30
        minLength: 3
        type: string
    required:
    - password
    - username
    type: object
...
  handler.UpdateProfileRequest:
    properties:
      avatar_url:
        type: string
      bio:
        type: string
    type: object
  handler.UpdateUsernameRequest:
    properties:
      username:
        maxLength: 30
        minLength: 3
        type: string
    required:
    - username
    type: object
...
  response.PageResponse:
    properties:
      has_more:
        type: boolean
      items: {}
      next_cursor:
        type: string
    type: object
  response.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
  response.TokenResponse:
    properties:
      access_token:
        type: string
      account_id:
        type: integer
      expires_in:
        type: integer
      refresh_token:
        type: string
    type: object
info:
  contact: {}
  description: 高性能短视频 Feed 流系统 API 文档
  termsOfService: http://swagger.io/terms/
  title: TideFlow 短视频 Feed 流 API
  version: v1
paths:
  /api/v1/auth/login:
    post:
      consumes:
      - application/json
      description: 使用用户名密码登录，返回access_token和refresh_token
      parameters:
      - description: 登录信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.TokenResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
      summary: 用户登录
      tags:
      - 认证
  /api/v1/auth/logout:
    post:
      description: 使当前access_token失效
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - OAuth2Password: []
      summary: 用户登出
      tags:
      - 认证
  /api/v1/auth/oauth2/token:
    post:
      consumes:
      - application/x-www-form-urlencoded
      description: Swagger UI登录专用端点，使用username/password换取access_token
      parameters:
      - description: 用户名
        in: formData
        name: username
        required: true
        type: string
      - description: 密码
        in: formData
        name: password
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.OAuth2TokenResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
      summary: OAuth2 Token (Swagger专用)
      tags:
      - 认证
  /api/v1/auth/password:
    post:
      consumes:
      - application/json
      description: 用户修改密码，需要提供旧密码验证
      parameters:
      - description: 密码修改信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.ChangePasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
      summary: 修改密码
      tags:
      - 认证
  /api/v1/auth/refresh:
    post:
      consumes:
      - application/json
      description: 使用refresh_token获取新的access_token和refresh_token
      parameters:
      - description: 刷新令牌
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.RefreshRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.TokenResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
      summary: 刷新令牌
      tags:
      - 认证
  /api/v1/auth/register:
    post:
      consumes:
      - application/json
      description: 注册新用户，返回access_token和refresh_token
      parameters:
      - description: 注册信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.TokenResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/response.Response'
      summary: 用户注册
      tags:
      - 认证
...
  /api/v1/users/{id}:
    get:
      description: 通过用户ID获取用户公开信息
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
      summary: 获取指定用户信息
      tags:
      - 用户资料
  /api/v1/users/{id}/followers:
    get:
      description: 获取指定用户的粉丝列表
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: integer
      - description: 游标
        in: query
        name: cursor
        type: string
      - default: 20
        description: 每页数量
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.PageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - OAuth2Password: []
      summary: 获取粉丝列表
      tags:
      - 社交关系
  /api/v1/users/{id}/following:
    get:
      description: 获取指定用户的关注列表
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: integer
      - description: 游标
        in: query
        name: cursor
        type: string
      - default: 20
        description: 每页数量
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.PageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - OAuth2Password: []
      summary: 获取关注列表
      tags:
      - 社交关系
  /api/v1/users/{id}/social-counts:
    get:
      description: 获取指定用户的关注数和粉丝数
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - OAuth2Password: []
      summary: 获取社交数量
      tags:
      - 社交关系
  /api/v1/users/{id}/videos:
    get:
      description: 获取指定用户发布的视频列表
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: integer
      - description: 游标
        in: query
        name: cursor
        type: string
      - default: 20
        description: 每页数量
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.PageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - OAuth2Password: []
      summary: 获取用户视频列表
      tags:
      - 用户资料
  /api/v1/users/{user_id}/follow:
    delete:
      description: 当前用户取消关注指定用户
      parameters:
      - description: 目标用户ID
        in: path
        name: user_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - OAuth2Password: []
      summary: 取消关注
      tags:
      - 社交关系
    post:
      description: 当前用户关注指定用户
      parameters:
      - description: 目标用户ID
        in: path
        name: user_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - OAuth2Password: []
      summary: 关注用户
      tags:
      - 社交关系
  /api/v1/users/me:
    get:
      description: 获取已登录用户的信息
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - OAuth2Password: []
      summary: 获取当前用户信息
      tags:
      - 用户资料
    put:
      consumes:
      - application/json
      description: 更新当前用户的头像和简介
      parameters:
      - description: 更新信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.UpdateProfileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - OAuth2Password: []
      summary: 更新个人资料
      tags:
      - 用户资料
  /api/v1/users/me/username:
    put:
      consumes:
      - application/json
      description: 修改当前用户的用户名
      parameters:
      - description: 新用户名
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.UpdateUsernameRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - OAuth2Password: []
      summary: 修改用户名
      tags:
      - 用户资料
  /api/v1/users/username/{username}:
    get:
      description: 通过用户名获取用户公开信息
      parameters:
      - description: 用户名
        in: path
        name: username
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
      summary: 通过用户名查找用户
      tags:
      - 用户资料
...
securityDefinitions:
  OAuth2Password:
    flow: password
    scopes:
      read: Grant read access
      write: Grant write access
    tokenUrl: /api/v1/auth/oauth2/token
    type: oauth2
swagger: "2.0"